refactor: scaffold the per-command tap registry
Restructure the monolithic commands.ts into a commands/ module: the
defineCommand authoring helper and TapCommandError move to definition.ts,
the health command to health.ts, and the registry itself to index.ts. The
TapManager gains the domain-failure (TapCommandError) catch and exec-args
coercion gets its own spec. This is the foundation the per-command modules
(specs, run, tests, commands) register into.
